fix: move getEntry tests from tests.test.ts to getEntry.test.ts
